+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #31 Cleaning up APIs deprecated since <= 0.25
 * #32 Update Jackson to 2.9.5
+* #33 Metrics renaming to follow naming convention
 
 ## 0.25